refactor: move metadata generators to separate files
